+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.0] - 2026-03-30
 
 ### Added
 
@@ -116,7 +116,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Git commit and annotated tag creation
 - `--dry-run` option to preview changes
 
-[Unreleased]: https://github.com/rodar-project/rodar_release/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/rodar-project/rodar_release/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/rodar-project/rodar_release/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/rodar-project/rodar_release/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/rodar-project/rodar_release/compare/v1.2.3...v1.3.0
 [1.2.3]: https://github.com/rodar-project/rodar_release/compare/v1.2.2...v1.2.3